improve stability of unit tests
TestMemoryUsage would sometimes manage to overrun it's loop bounds and
enter an infinite loop.

TestMultiThread would sometimes timeout if the machine was busy.

TestOutOfMemory could sometimes fail, and would then cause future tests
to fail because it held onto the memory.

FilePathReorderWrites was not deleting it's subsidiary memFS: data.

TestPreparedStatement::testDateTime8 had a typo.

The change to TestMVStore::testCacheSize I'm not sure about.