Document fetch first row(s) only
Since 1.4.177 H2 supports FETCH FIRST x ROW ONLY however this is not
reflected in the grammar documentation. This covers all the variations
of the syntax.

There are 3 `help.csv` but I believe I got the right one. Some of the
checked in files seem to have been automatically changed and I don't
know if I should have committed those as well.

Fixes #475