TimestampWithTimeZone can not be bound
Clients that are still on Java 7 and can not use OffsetDateTime need a
way to bind TimestampWithTimeZone. Unfortunately TimestampWithTimeZone
can currently not be used as a bind parameter
The obvious way to bind TimestampWithTimeZone would be to use
PreparedStatement.setObject(int, Object). This does not work because
DataTye#convertTo currently does not check for TimestampWithTimeZone.

This commit fixes this by adding a check and conversion in
DataTye#convertTo and also adding a unit test.