fix : The french messages are bad genereted (not contains DB message)
Error messages are translated in `org.h2.message.DbException` by using `MessageFormat.format` method.
But the french messages are not well generated by `MessageFormat`(`{X}` are  not replaced by `param` entry). 
That is normal, because they are not well formated : the single quote wasn't been escape
This PR solves this issue by escaping the french sentences for MessageFormat.

Java documentation of MessageFormat : https://docs.oracle.com/javase/8/docs/api/java/text/MessageFormat.html
An extract:
```
Within a String, a pair of single quotes can be used to quote any arbitrary characters except single quotes. For example, pattern string "'{0}'" represents string "{0}", not a FormatElement. A single quote itself must be represented by doubled single quotes '' throughout a String. For example, pattern string "'{''}'" is interpreted as a sequence of '{ (start of quoting and a left curly brace), '' (a single quote), and }' (a right curly brace and end of quoting), not '{' and '}' (quoted left and right curly braces): representing string "{'}", not "{}". 
```