Add support for custom Lucene analyzers
Lucene supports the concept of an "analyzer" - an abstraction
of text parsers suited for different type of text input. Some
of them are designed for natural languages, other for machine
text, some are customized for specific languages.

It would be nice if H2 provided the user an ability to set
a custom version of the analyzer, not only the standard one.
This would allow the user to tune the full-text search engine
for specific data.