Move documentation, add tests, refactor change logic
Add boolean flag to makeRegexpFlags that indicates whether the presence of a global flag should simply be ignored (i.e. when in PostreSQL mode) or whether it should throw an exception.

If the database is in PostgreSQL mode and the 'g' flag is not present, use the String.replaceFirst method instead of the H2 default String.replaceAll method when doing the regex replace.