Create Socket supports SO_TIMEOUT setting.
Socket SO_TIMEOUT is a core and important config in socket
communication. It helps developers to adjust read timeout
value according to bussiness and scenario they actually needs.

In Java, when SO_TIMEOUT set to a non-zero timeout, the read()
call associated with the socket stream will block util timeout is
reached, then throws a java.net.SocketTimeoutException. Otherwise,
when SO_TIMEOUT set to a zero timeout, the socket SO_TIMEOUT will be
interpreted as an infinite timeout. Zero timeout is the default setting.

Most of the time, use a default zero timeout with creating a new socket
is tranquil, But not everytime. I found that there are some cases will
make H2 connection hangs infinitly because of this problem.

Developer need this config in these scenes (It may be more than):

	- Long responding lead to socket connections accumulation.

	- Socket read block with infinite read timeout and finally not to be released leads to leak in NFS.
	  Associate readings:
	(https://access.redhat.com/solutions/3053801?spm=a2c4g.11186623.2.12.5ca22e518bdfck)
	(https://access.redhat.com/solutions/1427473?spm=a2c4g.11186623.2.13.5ca22e518bdfck)

	- Prevent the client socket from being blocked because the server's data is not returned(Application or Other Unknown underlying bugs).

So, one of the solution is make it allowed for H2 mantainer to config
socket SO_TIMEOUT in DB, avoid to hit the wall and be confused. The config
is named NETWORK_TIMEOUT and it can be configure in the H2 connection url.

Signed-off-by: Dyrone Teng <dyroneteng@gmail.com>